# server/swagger.yaml
swagger: "2.0"
info:
  title: Moving
  description: MovingApi
  version: 1.0.0
host: 192.168.31.49:8443
schemes:
  - https
basePath: /v1
paths:
...
  /items:
    post:
      summary: Save a item
      operationId: SaveItem
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/ItemData'
      responses:
        200:
          description: OK
          schema:
            type: string
        400:
          description: BadRequest
          schema:
            type: string
  /items/query:
    get:
      summary: Query items
      operationId: QueryItems
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: type
          in: query
          type: string
          required: true
        - name: tagName
          in: query
          type: string
        - name: keyword
          in: query
          type: string
          required: true
        - name: startIndex
          type: integer
          in: query
          required: true
        - name: fetchSize
          type: integer
          in: query
          required: true
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ItemData'
        400:
          description: BadRequest
          schema:
            type: string
...
definitions:
...
  ItemData:
    type: object
    properties:
      serverID:
        type: string
      boxCode:
        type: string
      code:
        type: string
      name:
        type: string
      description:
        type: string
      type:
        type: string
      tags:
        type: object
        additionalProperties:
          type: string
      count:
        type: integer
      props:
        type: object
        additionalProperties:
          type: string